simplified, tested, and added protection for update/removal
all calls are POST; all calls require standard message format with "key" to ensure the data is not changed or removed by unauthorized call to store/remove.

+++ UserPreferenceService/src/org/service/support/PSSupportStore.java
@@ -6,7 +6,7 @@ import elsu.common.*;
 import elsu.support.*;
 
 public class PSSupportStore {
-	
+
 	private static String _ENCRYPTIONKEYPATH = "application.encryption.key.key";
 	private static String _ENCRYPTIONVECTORPATH = "application.encryption.key.initVector";
 	private static String _STORAGEPATH = "application.framework.attributes.key.storage.repository";
@@ -14,12 +14,12 @@ public class PSSupportStore {
 	public static String array2JSON(String[] pKeys, String[] pValues) {
 		StringBuilder result = new StringBuilder();
 		Boolean firstTime = true;
-		
+
 		result.append("{");
 		if (pKeys.length == pValues.length) {
-			for (int i=0; i < pKeys.length; i++) {
+			for (int i = 0; i < pKeys.length; i++) {
 				result.append((firstTime ? "" : ",") + "\"" + pKeys[i] + "\": " + "\"" + pValues[i] + "\"");
-				if(firstTime) {
+				if (firstTime) {
 					firstTime = !firstTime;
 				}
 			}
@@ -32,78 +32,94 @@ public class PSSupportStore {
 	public static String map2JSON(HashMap<String, String> pMap) {
 		StringBuilder result = new StringBuilder();
 		Boolean firstTime = true;
-		
+
 		result.append("{");
-		for(String key : pMap.keySet()) {
+		for (String key : pMap.keySet()) {
 			result.append((firstTime ? "" : ",") + "\"" + key + "\": " + "\"" + pMap.get(key) + "\"");
-			if(firstTime) {
+			if (firstTime) {
 				firstTime = !firstTime;
 			}
 		}
 		result.append("}");
-		
+
 		return result.toString();
 	}
-	
-	public static String readPreference(GlobalSet pResources, ConfigLoader pConfig, String pId, String pKey) {
-		String result = "";
+
+	public static String readPreference(GlobalSet pResources, ConfigLoader pConfig, String pId, String pKey,
+			PreferenceType pPreference) {
+		String result = null;
 		String storePath = pConfig.getProperty(_STORAGEPATH).toString();
 		String fileName = storePath + (storePath.endsWith("/") ? "" : "/") + pId + "." + pKey;
-		
+
 		try {
 			// store the file
 			String content = FileUtils.readFile(fileName);
-
-			// encrypt the content
-			result = EncryptionUtils.decrypt(
-					pConfig.getProperty(PSSupportStore._ENCRYPTIONKEYPATH).toString(), 
-					pConfig.getProperty(PSSupportStore._ENCRYPTIONVECTORPATH).toString(), 
-					content);
+			
+			// decrypt the content
+			result = EncryptionUtils.decrypt(pPreference.getKey() + pConfig.getProperty(PSSupportStore._ENCRYPTIONKEYPATH).toString(),
+					pConfig.getProperty(PSSupportStore._ENCRYPTIONVECTORPATH).toString(), content);
+			
+			// compare the decrypted content to match original key
+			PreferenceType kpt = (PreferenceType) JsonXMLUtils.JSon2Object(result, PreferenceType.class);	
+			if (!kpt.getKey().equals(pPreference.getKey())) {
+				result = null;
+			}
 		} catch (Exception ex) {
 			pConfig.logError("readPreference(), \n" + ex.getMessage() + "\n" + ex.getStackTrace());
 		}
-		
+
 		return result;
 	}
-	
-	public static boolean storePreference(GlobalSet pResources, ConfigLoader pConfig, String pId, String pKey, String pContent) {
+
+	public static boolean storePreference(GlobalSet pResources, ConfigLoader pConfig, String pId, String pKey,
+			PreferenceType pPreference) {
 		boolean result = false;
 		String storePath = pConfig.getProperty(_STORAGEPATH).toString();
 		String fileName = storePath + (storePath.endsWith("/") ? "" : "/") + pId + "." + pKey;
-		
+
 		try {
+			// if file does exists, then validate it for ownership
+			if (FileUtils.fileExists(storePath, (pId + "." + pKey)) > 0) {
+				if (readPreference(pResources, pConfig, pId, pKey, pPreference) == null) {
+					return false;
+				}
+			}
+			
 			// encrypt the content
 			String encryptedData = EncryptionUtils.encrypt(
-					pConfig.getProperty(PSSupportStore._ENCRYPTIONKEYPATH).toString(), 
+					pPreference.getKey() + pConfig.getProperty(PSSupportStore._ENCRYPTIONKEYPATH).toString(),
 					pConfig.getProperty(PSSupportStore._ENCRYPTIONVECTORPATH).toString(), 
-					pContent);
+					JsonXMLUtils.Object2JSon(pPreference));
 
 			// store the file
 			FileUtils.writeFile(fileName, encryptedData, true);
-			
+
 			// update result
 			result = true;
 		} catch (Exception ex) {
 			pConfig.logError("storePreference(), \n" + ex.getMessage() + "\n" + ex.getStackTrace());
 		}
-		
-		result = true;
+
 		return result;
 	}
-	
-	public static boolean deletePreference(GlobalSet pResources, ConfigLoader pConfig, String pId, String pKey) {
+
+	public static boolean deletePreference(GlobalSet pResources, ConfigLoader pConfig, String pId, String pKey,
+			PreferenceType pPreference) {
 		boolean result = false;
 		String storePath = pConfig.getProperty(_STORAGEPATH).toString();
 		String fileName = storePath + (storePath.endsWith("/") ? "" : "/") + pId + "." + pKey;
-		
+
 		try {
-			// remove the file
-			FileUtils.deleteFile(fileName);
-			result = true;
+			// check to make sure the preference is owned by the user
+			if (readPreference(pResources, pConfig, pId, pKey, pPreference) != null) {
+				// remove the file
+				FileUtils.deleteFile(fileName);
+				result = true;
+			}
 		} catch (Exception ex) {
 			pConfig.logError("deletePreference(), \n" + ex.getMessage() + "\n" + ex.getStackTrace());
 		}
-		
+
 		return result;
 	}
 }
